fix /blogs
